// src/types/enrollment.ts
export interface Semester {
  id: string;
  semesterName: string; // e.g., "1st Sem 2025–2026"
  academicYear: string; // e.g., "2025-2026"
  semesterType: '1st Semester' | '2nd Semester' | 'Summer';
  startDate: string;
  endDate: string;
  enrollmentStartDate: string;
  enrollmentEndDate: string;
  status: 'Upcoming' | 'Active' | 'Completed' | 'Cancelled';
  dateCreated: string;
  dateUpdated: string;
}